requirements.txt
easyocr==1.6.2
matplotlib==3.6.2
opencv-python-headless==4.5.4.60
Pillow==9.5.0
PyAutoGUI==0.9.53
pynput==1.7.6



